Separated out the loaders into individual files and started on pneumonia loader
