+++ aggregators/FedMGDAPlus.py
@@ -142,7 +142,7 @@ class FedMGDAPlusAggregator(Aggregator):
                 # Increasing / Decreasing LR each global round
                 for g in self.lambdatOpt.param_groups:
                     g["lr"] *= 0.9**size
-                    print(f"New LR: {g["lr"]}")
+                    print(f"New LR: {g['lr']}")
 
                     std_multiplier *= 1.05**size
                     print(f"New std: {std_multiplier}")
@@ -151,7 +151,7 @@ class FedMGDAPlusAggregator(Aggregator):
                 # Increasing / Decreasing LR each global round
                 for g in self.lambdatOpt.param_groups:
                     g["lr"] *= 1.3
-                    print(f"New LR: {g["lr"]}")
+                    print(f"New LR: {g['lr']}")
 
                     std_multiplier /= 1.05
                     print(f"New std: {std_multiplier}")
